Remove old JSON Reader and Decoder (#3610)
