+++ arrow-json/src/lib.rs
@@ -25,10 +25,18 @@
 pub mod reader;
 pub mod writer;
 
-mod raw;
+#[doc(hidden)]
+#[deprecated(note = "Use Decoder")]
+pub type RawDecoder = reader::Decoder;
 
-pub use self::raw::{RawDecoder, RawReader, RawReaderBuilder};
-#[allow(deprecated)]
+#[doc(hidden)]
+#[deprecated(note = "Use Reader")]
+pub type RawReader<R> = Reader<R>;
+
+#[doc(hidden)]
+#[deprecated(note = "Use ReaderBuilder")]
+pub type RawReaderBuilder = ReaderBuilder;
+
 pub use self::reader::{Reader, ReaderBuilder};
 pub use self::writer::{ArrayWriter, LineDelimitedWriter, Writer};
 use half::f16;
