+++ arrow-json/src/reader/serializer.rs
@@ -15,7 +15,7 @@
 // specific language governing permissions and limitations
 // under the License.
 
-use crate::raw::tape::TapeElement;
+use crate::reader::tape::TapeElement;
 use lexical_core::FormattedSize;
 use serde::ser::{
     Impossible, SerializeMap, SerializeSeq, SerializeStruct, SerializeTuple,
