+++ arrow-json/src/reader/tape.rs
@@ -15,7 +15,7 @@
 // specific language governing permissions and limitations
 // under the License.
 
-use crate::raw::serializer::TapeSerializer;
+use crate::reader::serializer::TapeSerializer;
 use arrow_schema::ArrowError;
 use serde::Serialize;
 use std::fmt::{Display, Formatter};
