+++ parquet/src/arrow/arrow_writer/levels.rs
@@ -1145,7 +1145,7 @@ mod tests {
             false,
         );
         let schema = Arc::new(Schema::new(vec![stocks_field]));
-        let builder = arrow::json::RawReaderBuilder::new(schema).with_batch_size(64);
+        let builder = arrow::json::ReaderBuilder::new(schema).with_batch_size(64);
         let mut reader = builder.build(std::io::Cursor::new(json_content)).unwrap();
 
         let batch = reader.next().unwrap().unwrap();
